cast
```
 In file included from C:\projects\libssh2\_bld\src\CMakeFiles\libssh2_shared.dir\Unity\unity_0_c.c:73:
C:/projects/libssh2/src/session.c(1341,12): error : implicit conversion loses integer precision: 'libssh2_int64_t' (aka 'long long') to 'long' [-Werror,-Wshorten-64-to-32] [C:\projects\libssh2\_bld\src\libssh2_shared.vcxproj]
C:/projects/libssh2/src\libssh2_priv.h(307,41): message : expanded from macro 'ssh2_timediff_to_sec' [C:\projects\libssh2\_bld\src\libssh2_shared.vcxproj]
  In file included from C:\projects\libssh2\_bld\src\CMakeFiles\libssh2_static.dir\Unity\unity_0_c.c:73:
C:/projects/libssh2/src/session.c(1341,12): error : implicit conversion loses integer precision: 'libssh2_int64_t' (aka 'long long') to 'long' [-Werror,-Wshorten-64-to-32] [C:\projects\libssh2\_bld\src\libssh2_static.vcxproj]
C:/projects/libssh2/src\libssh2_priv.h(307,41): message : expanded from macro 'ssh2_timediff_to_sec' [C:\projects\libssh2\_bld\src\libssh2_static.vcxproj]
```
https://ci.appveyor.com/project/libssh2org/libssh2/builds/54396600/job/7s8e8dp42nvu2us6?fullLog=true

+++ src/session.c
@@ -1385,7 +1385,7 @@ long libssh2_session_get_read_timeout(LIBSSH2_SESSION *session)
     if(!session)
         return 0;
 
-    return ssh2_timediff_to_sec(session->packet_read_timeout);
+    return (long)ssh2_timediff_to_sec(session->packet_read_timeout);
 }
 
 #ifndef LIBSSH2_NO_DEPRECATED
